requirements.txt
einops==0.3.2
numpy==1.19.2
opencv-python-headless==4.5.3.56
Pillow==8.4.0
tensorflow==2.7.0
timm==0.4.12
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
